+++ testing/chain.go
@@ -42,6 +42,10 @@ type SenderAccount struct {
 	SenderAccount authtypes.AccountI
 }
 
+const (
+	DefaultGenesisAccBalance = "10000000000000000000"
+)
+
 // TestChain is a testing struct that wraps a simapp with the last TM Header, the current ABCI
 // header and the validators of the TestChain. It also contains a field called ChainID. This
 // is the clientID that *other* chains use to refer to this TestChain. The SenderAccount
@@ -104,7 +108,7 @@ func NewTestChainWithValSet(t *testing.T, coord *Coordinator, chainID string, va
 	for i := 0; i < MaxAccounts; i++ {
 		senderPrivKey := secp256k1.GenPrivKey()
 		acc := authtypes.NewBaseAccount(senderPrivKey.PubKey().Address().Bytes(), senderPrivKey.PubKey(), uint64(i), 0)
-		amount, ok := sdk.NewIntFromString("10000000000000000000")
+		amount, ok := sdk.NewIntFromString(DefaultGenesisAccBalance)
 		require.True(t, ok)
 
 		// add sender account